Got rid of top limit switch height calibration.
We had a problem where the elevator smashed down to the home position and broke the motor. The command we ran should have had the elevator go up, so we think that it got reset to thinking it was at the top. Removing the top calibration will make the code safer.